Hide internals of the store view implementation
